+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 0.22.6 (2023-08-02)
 
 * [Feature] Add `mermaid` backend to `ploomber plot` (by [@TomArm](https://github.com/TomArm))
+* [Fix] Fixes incompatibility in `ploomber install` when using an old version of `ploomber-core`
 
 ## 0.22.5 (2023-07-27)